Fix SubscribeOn handling of concurrent scheduling
The susbscriber was calling Wait on a concurrent scheduler when its own
Wait method was called. If you called Wait on a subscription it would
acually wait for all concurrent tasks (goroutines) to terminate.

When the subscriber unsubscribes it will call cancel on all tasks that 
were scheduled under its subscription.

There is no need to Wait for all tasks to finish. So we set OnWait to
nil so it will wait for the subscriber to be no longer subscribed.